requirements.txt
Django==4.2.6
djangorestframework==3.14.0
python-decouple==3.8
python_jose==3.3.0
